requirements.txt
psycopg[binary]
anthropic>=0.34.0
python-dotenv
fastapi
uvicorn[standard]
google-cloud-bigquery>=3.0.0
